+++ CHANGELOG.md
@@ -7,13 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Added theme manager to get and set the window theme either explicit or by system.
+- Added SystemThemeProvider to read the current windows theme.
+- Added AccentColorProvider to read the current windows accent colors.
+- Added theme manager to set the window theme either explicit or by system.
 - Added the possibility to attach the requested theme to each window to have it applied automatically.
-- Added ThemedWindow object for more easy usage.
-- Added AccentColorProvider to read the windows accent colors.
 - Added AccentBrush markup extension to use the accent colors on UI controls directly.
 - Added AccentColor markup extension to use the accent colors within color brushes.
 - Added color set change observer to get informed when the theme or accent color on windows got changed.
+- Added ThemedWindow object for more easy usage of themes.
 ### Supported .Net Versions
 - .Net Core 3.0
 - .Net Framework 4.5